requirements.txt
requests
psycopg2-binary==2.7.6.1
